separate l1track types
